Refactor leveldb persistence, split into mutliple files
